Day 1 Solution Upload
First implementation was pretty hacky due to issues with dynamic ranges and excel on iPad. Solution uploaded is optimized to one helper column and then two cells to solve part 1 and 2. An assumption is made that no two elves have overlapping maximum values. The first run I did accounted for this but it did not seem to be an issue for my data set. The excel file is very large due to excel for iPad handling of dynamic ranges.